requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
finstat==0.11.1
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.1
ply==3.11
Pyomo==6.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
xlwings==0.27.2
